fw: Improve portability of "struct cr_args"
msp430-libc does not typedef ssize_t (the singed version of size_t). It does
define _ssize_t as a macro in sys/_types.h; but this seems hackish.

The use case here is fine with the interger range of "int". So, let's
use that.

+++ remote-fw/protocol.h
@@ -25,9 +25,9 @@ enum cr_requests {
 };
 
 struct cr_args {
-    size_t min;
+    unsigned int min;
     /** If -1, there is no maximum */
-    ssize_t max;
+    int max;
 };
 
 struct cr_word {
